Renamed module folder names to lowercase.
