+++ include/dsannotation/serialization/IManifestWriter.h
