+++ include/dsannotation/serialization/JsonManifestWriter.h
