+++ src/core/Component.cpp
@@ -1,4 +1,4 @@
-#include "DSAnnotation/Core/Component.h"
+#include "dsannotation/core/Component.h"
 
 #include <utility>
