+++ src/core/Reference.cpp
@@ -1,4 +1,4 @@
-#include "DSAnnotation/Core/Reference.h"
+#include "dsannotation/core/Reference.h"
 
 namespace dsannotation::core {
