+++ src/parsing/AnnotationValidator.cpp
@@ -1,4 +1,4 @@
-#include "DSAnnotation/Parsing/AnnotationValidator.h"
+#include "dsannotation/parsing/AnnotationValidator.h"
 #include <algorithm>
 #include <stack>
 #include <cctype>
