+++ src/serialization/JsonManifestWriter.cpp
@@ -1,4 +1,4 @@
-#include "DSAnnotation/Serialization/JsonManifestWriter.h"
+#include "dsannotation/serialization/JsonManifestWriter.h"
 
 #include <exception>
