+++ src/support/ErrorReporter.cpp
@@ -1,4 +1,4 @@
-#include "DSAnnotation/Support/ErrorReporter.h"
+#include "dsannotation/support/ErrorReporter.h"
 
 #include <iostream>
 #include <sstream>
